feat(cli): implement download and package management commands
