+++ cmd/ppkgmgr/root.go
@@ -1,0 +1,25 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+func newRootCmd(downloader downloadFunc) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:           "ppkgmgr",
+		Short:         "Private package manager CLI",
+		SilenceErrors: true,
+		SilenceUsage:  true,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			fmt.Fprintln(cmd.ErrOrStderr(), "require subcommand (run 'ppkgmgr help' for usage)")
+			return cliError{code: 1}
+		},
+	}
+
+	cmd.AddCommand(newDownloadCmd(downloader))
+	cmd.AddCommand(newPkgCmd())
+	cmd.AddCommand(newVersionCmd())
+	return cmd
+}
