+++ go.mod
@@ -2,9 +2,14 @@ module ppkgmgr
 
 go 1.24
 
-require gopkg.in/yaml.v3 v3.0.1
+require (
+	github.com/spf13/cobra v1.8.1
+	github.com/zeebo/blake3 v0.2.4
+	gopkg.in/yaml.v3 v3.0.1
+)
 
 require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/klauspost/cpuid/v2 v2.0.12 // indirect
-	github.com/zeebo/blake3 v0.2.4 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 )
